--- yak_ros/package.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0"?>
-<package format="3">
-  <name>yak_ros</name>
-  <version>0.1.0</version>
-  <description>An example ROS node for yak</description>
-  <author email="Joseph.Schornak@SwRI.org">Joseph Schornak</author>
-  <author>Jonathan Meyer</author>
-  <maintainer email="Joseph.Schornak@SwRI.org">Joseph Schornak</maintainer>
-  <license>Apache 2.0</license>
-
-  <buildtool_depend condition="$ROS_VERSION == 1">catkin</buildtool_depend>
-  <buildtool_depend condition="$ROS_VERSION == 2">ament_cmake</buildtool_depend>
-
-  <depend condition="$ROS_VERSION == 1">roscpp</depend>
-  <depend condition="$ROS_VERSION == 2">rclcpp</depend>
-
-
-  <depend>libpcl-all-dev</depend>
-  <depend>eigen</depend>
-  <depend>yak</depend>
-  <depend>yak_ros_msgs</depend>
-
-  <depend>sensor_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>std_msgs</depend>
-  <depend>std_srvs</depend>
-  <depend>visualization_msgs</depend>
-
-  <depend>tf2</depend>
-  <depend>tf2_ros</depend>
-  <depend>tf2_eigen</depend>
-
-  <depend>pcl_ros</depend>
-  <depend>cv_bridge</depend>
-  <depend>image_transport</depend>
-  <depend condition="$BUILD_DEMO == True">gl_depth_sim</depend>
-
-  <export>
-    <build_type condition="$ROS_VERSION == 1">catkin</build_type>
-    <build_type condition="$ROS_VERSION == 2">ament_cmake</build_type>
-  </export>
-
-</package>
