+++ yak_ros2/package.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0"?>
+<package format="3">
+  <name>yak_ros2</name>
+  <version>0.1.0</version>
+  <description>An example ROS2 node for yak</description>
+  <author email="Joseph.Schornak@SwRI.org">Joseph Schornak</author>
+  <author>Jonathan Meyer</author>
+  <maintainer email="Joseph.Schornak@SwRI.org">Joseph Schornak</maintainer>
+  <license>Apache 2.0</license>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+
+  <depend>cv_bridge</depend>
+  <depend>eigen</depend>
+  <depend>geometry_msgs</depend>
+  <depend condition="$BUILD_DEMO == True">gl_depth_sim</depend>
+  <depend>image_transport</depend>
+  <depend>libpcl-all-dev</depend>
+  <depend>pcl_ros</depend>
+  <depend>rclcpp</depend>
+  <depend>sensor_msgs</depend>
+  <depend>std_msgs</depend>
+  <depend>std_srvs</depend>
+  <depend>tf2</depend>
+  <depend>tf2_eigen</depend>
+  <depend>tf2_ros</depend>
+  <depend>visualization_msgs</depend>
+  <depend>yak</depend>
+  <depend>yak_ros_msgs</depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
