+++ yak_ros2_msgs/package.xml
@@ -1,18 +1,15 @@
 <?xml version="1.0"?>
 <package format="3">
-  <name>yak_ros_msgs</name>
+  <name>yak_ros2_msgs</name>
   <version>0.0.0</version>
-  <description>The yak_ros_msgs package</description>
+  <description>The yak_ros2_msgs package</description>
 
   <maintainer email="Joseph.Schornak@SwRI.org">Joseph Schornak</maintainer>
   <license>Apache 2.0</license>
 
+  <buildtool_depend>ament_cmake</buildtool_depend>
 
-  <buildtool_depend condition="$ROS_VERSION == 1">catkin</buildtool_depend>
-  <buildtool_depend condition="$ROS_VERSION == 2">ament_cmake</buildtool_depend>
-
-  <depend condition="$ROS_VERSION == 1">roscpp</depend>
-  <depend condition="$ROS_VERSION == 2">rclcpp</depend>
+  <depend>rclcpp</depend>
 
   <build_depend>geometry_msgs</build_depend>
   <build_export_depend>geometry_msgs</build_export_depend>
@@ -24,9 +21,9 @@
   <build_depend>std_msgs</build_depend>
   <build_export_depend>std_msgs</build_export_depend>
 
-
   <export>
-    <build_type condition="$ROS_VERSION == 1">catkin</build_type>
-    <build_type condition="$ROS_VERSION == 2">ament_cmake</build_type>
+    <build_type>ament_cmake</build_type>
   </export>
+
+  <member_of_group>rosidl_interface_packages</member_of_group>
 </package>
