--- yak_ros_msgs/CMakeLists.txt
@@ -1,65 +1,0 @@
-cmake_minimum_required(VERSION 2.8.3)
-project(yak_ros_msgs)
-
-find_package(catkin REQUIRED COMPONENTS
-  geometry_msgs
-  message_generation
-  std_msgs
-)
-
-
-################################################
-## Declare ROS messages, services and actions ##
-################################################
-
-# Generate messages in the 'msg' folder
-add_message_files(
-  FILES
-  KinFuParams.msg
-)
-
-# Generate services in the 'srv' folder
- add_service_files(
-   FILES
-   GenerateMesh.srv
-   UpdateKinFuParams.srv
- )
-
-## Generate actions in the 'action' folder
-# add_action_files(
-#   FILES
-#   Action1.action
-#   Action2.action
-# )
-
-# Generate added messages and services with any dependencies listed here
-generate_messages(
-  DEPENDENCIES
-    geometry_msgs
-    std_msgs
-)
-
-
-###################################
-## catkin specific configuration ##
-###################################
-catkin_package(
-#  INCLUDE_DIRS include
-#  LIBRARIES yak_ros_msgs
-  CATKIN_DEPENDS
-    geometry_msgs
-    message_runtime
-    std_msgs
-#  DEPENDS system_lib
-)
-
-###########
-## Build ##
-###########
-
-## Specify additional locations of header files
-## Your package locations should be listed before other locations
-include_directories(
-# include
-  ${catkin_INCLUDE_DIRS}
-)
